pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "motorways"
version = "0.1.0"
description = "macOS screen-capture + input-control RL player for Mini Motorways"
authors = [{name = "AI Development", email = "dev@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "pyautogui>=0.9.50",
    "pyobjc-framework-Cocoa>=9.0",
    "pyobjc-framework-Quartz>=9.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "stable-baselines3>=2.0.0",
    "torch>=1.10.0",
    "gymnasium>=0.26.0",
    "pathlib",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
motorways = "motorways.app.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"